Removed XMPIterator.kt interface as there is only one implementation.
